now you can resize a matrix from the gui :3
you just can't really tell ;D
